requirements.txt
torch>=1.7.0
numpy
pandas
tqdm
scikit-learn
PyYAML
EduCDM
